move parallelNelderMead to estimation
